polaris: shims: rename libpiex_shim to avoid conflict
The current name has conflict with the one in hardware/custom/compat.
I chose not to use that shim because mangling rules are not part of
the c++ standard. A shim should always use the exposed symbol ABI
directly instead of relying on the compiler to generate correct
mangled symbols from c++ code.

Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>